Adding support for multi-threading

+++ targetDevice/src/track_laser.py
@@ -11,6 +11,7 @@ import time
 import json
 import RPi.GPIO as gpio
 import random
+import threading
 
 '''
 Uses a camera to track a laser pointer and instruct it, via dictation,
@@ -19,10 +20,10 @@ to move toward a destination.
 class LaserTracker(object):
 
     def __init__(self):
-        self.gpio_pins = [18, 23]
+        self.gpio_pins = [18, 23, 24]
         # How often we'll dictate commands
-        self.command_interval = 2 #seconds
-        self.hit_radius_px = 60 # pixels
+        self.command_interval = 1.5 #seconds
+        self.hit_radius_px = 70 # pixels
         self.defaultRegion = 'us-east-1'
         self.defaultPollyEndpoint = 'https://polly.us-east-1.amazonaws.com'
         gpio.setmode(gpio.BCM)
@@ -56,29 +57,29 @@ class LaserTracker(object):
     '''
     def speak(self, polly, text, format='ogg_vorbis', voice='Joanna'):
         now = time.time()
-        filename = 'tmp.ogg'
-        #filename = text.replace(' ', '_') + '.ogg'
+        #filename = 'tmp.ogg'
+        filename = text.replace(' ', '_') + '.ogg'
         # look for cached result
-        #if os.path.isfile(filename):
-        #    print("Previous audio file found. Playing: " + filename)
-        #    os.system('omxplayer ' + filename)
-        #else:
+        if os.path.isfile(filename):
+            print("Previous audio file found. Playing: " + filename)
+            os.system('omxplayer ' + filename)
+        else:
             # No cached result, so fetch from Polly
-        #print("No previous result for audio filename: " + filename + ". Calling Polly...")
-        resp = polly.synthesize_speech(OutputFormat=format, Text=text, VoiceId=voice)
-        with closing(resp["AudioStream"]) as stream:
-            # mp3 files were playing the with end clipped via omxplayer.
-            soundfile = open(filename, 'wb')
-            soundfile.write(stream.read())
-            soundfile.flush()
-            soundfile.close()
-            os.system('omxplayer ' + filename + ' > /dev/null')
+            print("No previous result for audio filename: " + filename + ". Calling Polly...")
+            resp = polly.synthesize_speech(OutputFormat=format, Text=text, VoiceId=voice)
+            with closing(resp["AudioStream"]) as stream:
+                # mp3 files were playing the with end clipped via omxplayer.
+                soundfile = open(filename, 'wb')
+                soundfile.write(stream.read())
+                soundfile.flush()
+                soundfile.close()
+                os.system('omxplayer ' + filename + ' > /dev/null')
 
         print("Spoke audio in " + str(time.time() - now) + " seconds")
 
     '''
     '''
-    def send_to_lex(self, polly, lex, text='Hello world', botName='testing', botAlias='test_alias', userId='targetingDefault', voice='Joanna', \
+    def send_to_lex(self, polly_client, lex_client, text='Hello world', botName='testing', botAlias='test_alias', userId='targetingDefault', voice='Joanna', \
                           lexContentType='audio/x-l16; sample-rate=16000; channel-count=1'):
         
         lex_parsed = False
@@ -86,10 +87,10 @@ class LaserTracker(object):
         while not lex_parsed:
             start = time.time()
             print("Calling Polly for command " + text)
-            resp = polly.synthesize_speech(OutputFormat='pcm', SampleRate='16000', Text=text, VoiceId=voice)
+            resp = polly_client.synthesize_speech(OutputFormat='pcm', SampleRate='16000', Text=text, VoiceId=voice)
             print("Sending Polly response to Lex for command " + text)
             with closing(resp["AudioStream"]) as stream:
-                lex_resp = lex.post_content(botName=botName, botAlias=botAlias, userId=userId, contentType=lexContentType, inputStream=stream.read())
+                lex_resp = lex_client.post_content(botName=botName, botAlias=botAlias, userId=userId, contentType=lexContentType, inputStream=stream.read())
                 print("Fetched audio and posted to Lex in " + str(time.time() - start) + " seconds")
                 
                 if lex_resp['dialogState'] == 'ElicitIntent':
@@ -106,7 +107,7 @@ class LaserTracker(object):
         # define the lower and upper boundaries
         # in the HSV color space
         lower_red = np.array([166, 31, 122])
-        upper_red = np.array([250, 150, 255])
+        upper_red = np.array([250, 250, 255])
         lower_green = np.array([40, 15, 140])
         upper_green = np.array([72, 255, 255])
         
@@ -186,14 +187,16 @@ class LaserTracker(object):
         
         # initialize the camera and grab a reference to the raw camera capture
         # with-block ensures camera.close() is called upon exit.
-        with PiCamera(resolution = (640, 480),framerate = 15) as camera:
+        #with PiCamera(resolution = (640, 480),framerate = 15) as camera:
+        with PiCamera(resolution = (1280, 960),framerate = 15) as camera:
             try:
-                camera.hflip = True
-                camera.vflip = True
+                #camera.hflip = True
+                #camera.vflip = True
 
                 # allow the camera to warmup
                 time.sleep(0.1)
-                rawCapture = PiRGBArray(camera, size=(640, 480))
+                #rawCapture = PiRGBArray(camera, size=(640, 480))
+                rawCapture = PiRGBArray(camera, size=(1280, 960))
                 
                 # initialize the next time to dictate a command to now
                 next_command_time = time.time()
@@ -251,8 +254,15 @@ class LaserTracker(object):
                                     command = "move down"
                             
                             print(command)
-                            self.speak(polly, command)
-                            self.send_to_lex(polly=polly, lex=lex, text=command)
+                            polly_thread = threading.Thread(target=self.speak, args=(polly, command))
+                            lex_thread = threading.Thread(target=self.send_to_lex, args=(polly, lex, command))
+                            polly_thread.start()
+                            lex_thread.start()
+                            polly_thread.join()
+                            lex_thread.join()
+                            #self.speak(polly, command)
+                            #self.send_to_lex(polly_client=polly, lex_client=lex, text=command)
+                            
                             next_command_time = time.time() + self.command_interval
                             prev_loc_diff = location_difference
